An initial setup with modified build process to integrate adonis' edge.js templating for panel dev.
In this initial commit the master view is pre-compiled during the build with an additional watcher added to watch changes to the contents of a new 'views' folder. The compiled template is written to the 'dist/chrome' folder as panel.html, so the original copy step for panel.html has been removed.